style: move files related to backend into backend/
